fix: modify findById controller to use Long and not Integer

+++ src/main/java/me/cerrato/w2m/spaceships/api/controllers/ApiController.java
@@ -72,7 +72,7 @@ public class ApiController {
                     })
     })
     @GetMapping("/{id}")
-    public Spaceship findById(@PathVariable("id") Integer id) throws EntityDontExistException {
+    public Spaceship findById(@PathVariable("id") Long id) throws EntityDontExistException {
         SpaceshipEntity entity = service.findById(id);
         return conversionService.convert(entity, Spaceship.class);
     }
